requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.7
django-crispy-forms==2.1
django-localflavor==4.0
djangorestframework==3.14.0
Levenshtein==0.23.0
Pillow==10.1.0
python-barcode==0.15.1
python-Levenshtein==0.23.0
python-stdnum==1.19
pytz==2023.3.post1
rapidfuzz==3.5.2
sqlparse==0.4.4
tzdata==2023.3
